Add files via upload
The program first takes input from the user using the scanf() function and stores it in the variable n. Then the program uses a for loop to iterate through the range of numbers from 1 to n. The for loop uses the variable i as the loop counter and initializes it to 1. The loop continues until i is less than or equal to n. In each iteration of the loop, the current value of i is printed using the printf() function, and then i is incremented by 1 using the i++ statement. The \n at the end of the printf() statement is used to print each number on a new line.